MAA-LLM/maa-django-backend/requirements.txt
Django==5.1
djangorestframework==3.15
langchain==0.2.16
langchain-community==0.2.16
langchain-groq==0.1.9
llama-index-llms-groq==0.3.1
llama-index==0.11.20
llama-index-llms-ollama==0.3.2
llama-index-embeddings-huggingface==0.3.1
python-dotenv==1.0
pydantic==2.9
djangorestframework-simplejwt==5.3
django-cors-headers==4.4.0
